Add spiking neuron calibration
Add a "wrapper" function which calls parameter calibrations
in a way suitable for normal LIF neuron usage.
The calibrations use the CADCs, the MADC and rate counters
for different parameters.

Further improvements can target the following points:
- Getting rid of global configurations in c/14462.
- Member functions which takes a `builder` as an input return the
altered `builder`. This behavior might not be anticipated and should be
fixed (issue 3952).
- The test should be extended to assert the calibration targets are
reached, currently it is only checked if some spikes can be observed
(issue 3965).

Change-Id: Ieea6027afc72d1f2bc608c2922b37c09698796c1