Add MADC-based reference potential calib
The synaptic input reference potentials can be calibrated
by ensuring no current flows on the membrane (while the leak
is disabled).
This measurement can be performed with the MADC and may be more
precise than using the fast CADC method.

The fast CADC method will still be used as the default method in
the HAGEN mode calibration.

Change-Id: Icef3a86bc50411399f5fa6ea27a4bc5599693563