Use lola neurons in hagen and spiking results
Transitions the calib result classes to lola neurons
and a dict of global configurations.

We now save all the relevant parameters in the calib results, even ones
with noisy defaults, so we ensure the distribution of noise is the same
when applying the calib. This is important because of CapMem crosstalk.

Change-Id: I8567a0346ef84b7bfd40ba6c6000bc1de29cb026