Improve membrane time constant calib
Leak division or multiplication is now selected based on the individual
neurons' behavior and not based on a simple global threshold.
A new stimulation method, using offset currents instead of resets,
is used for target membrane time constants that are not too short.

Possible limitations:
The original neuron config is not restored after the offset-based
membrane time constant has run. Instead, the config used (and altered
due to leak division) during the calib remains on the chip. We provide
some reasoning for that in comments in the code.

Change-Id: I0264970df831c9765820d41578b01486c916bd4b