Add neuron threshold calibration
Add calibration which sets the threshold potential slightly above
the current leak potential. This allows spiking with only small
synaptic inputs.

Possible limitations:
- The leak potential has to reach the desired threshold, which means the
threshold can not be calibrated above some 0.8 to 1.0 V (as the leak
potential does not reach higher). This is solved in c/12859 by sampling
the membrane potential with the MADC or CADC to measure the threshold,
instead of using the spike counter here.
- The timing of spike counter resets is aligned with the timing of reads
by executing two resets successively. This may not be exact or may break
in the future. To improve this, one could break up the measurement and
individually execute a reset-wait-read scheme for each neuron.
- The assertion criteria in the test could be stronger. They are
improved in c/11447 for HICANN-X v2.

Further improvements can target the following points:
- Getting rid of global configurations in c/14462.
- one member function which takes a `builder` as an input returns the
altered `builder`. This behavior might not be anticipated and should be
fixed (issue 3952).

Change-Id: I91619b30f62ba308d6276e3cccb2bd0bcb238ba2