Add MADC- and CADC-based neuron threshold calib
Instead of relying on a high leak potential and the spike counters,
two new calibrations use a constant current onto the membrane
and record the voltage continuously. The maximum measured
voltage is assumed to be (just below) the threshold.

The spiking neuron calibration now uses the `ThresholdCalibCADC`
class to calibrate the threshold, which allows for a larger range of
threshold potentials and better accuracy, as the previous method
was limited by the available range of the leak potential.

Change-Id: Ie251618bfc03efa72479400bb917f60309edad3d