Setting of the 'holdoff period'
Allow to choose a target for the holdoff time.

Depending on the chosen refractory and holdoff time the neurons are
assigned to one of the two refractory clocks, clock scalers are
determined and the counter values for both times are determined.

In addition to the 'calibration routine' an example is provided how to
perform a calibration for neurons which are able to generate plateau
potentials.

Currently the same clock scales are chosen for all CommonNeuronBackends.
In future scalers could be chosen independently to increase the accuracy
of the refractory and holdoff time.

Change-Id: If4c73171950c6438750a66dbd142a7810b35abae