Add hagen synaptic input integration default calib
Along the default hagen and spiking calib, we deploy a new calibration
result that contains necessary configurations for integration on the
synaptic input lines, as opposed to neurons' membranes.

Possible limitations:
* Even when using integration on the synaptic inputs, we currently rely
on calibrated neurons during the synapse driver calibration. Since the
neuron calib result is available anyway now, we store it with the
HagenSyninCalibrationResult, which is now an extension of the (usual)
HagenCalibrationResult. If we develop a synapse driver calibration based
on the synaptic input lines, not necessitating calibrated neurons, we
will drop the neuron calib result from the HagenSyninCalibrationResult.
* We currently only test whether the calibration runs and generates a
result. A quantitative analysis of the results will be added
later, in changeset 16177.

Change-Id: Ic5f93a9e2551ea9b7749113b1d42b140f417220a