Fix repeated triggering of the MADC
The MADC will immediately start another measurement if the
`start_recording` flag is still active. If after triggering
a measurement, a following wait exceeds the sampling time
of the MADC, it will trigger multiple times.

We fix this possible bug by disabling the `start_recording` flag
immediately after setting it, so that there is no (possibly long) wait
in between.

Change-Id: I8f49990bfcf58fa4861fff2ee42754c5647ba631