Restore parameters after using offset currents
In some calibrations, we use offset currents to the neurons' membranes.
We did not set the repsective CapMem cells back to zero in the postlude,
but this may result in crosstalk. We set them back to zero now.

Also, the neuron config is now restored in the ThresholdCalibCADC,
where it is altered during the calibration.

Change-Id: I127e1b31c6be07d12a47b84f2f2d4cfc7cf0ec3a