Fix synapse DAC bias calib prelude
When finding the appropriate number of enabled synapse rows during the
synapse DAC bias calib prelude, the search did not continue if the
target was not convertible to int (like 55.5) or exactly at the upper
end of the dynamic range. Instead, an exeption was raised if the
target was exactly 60.

Here, we fix the update rule of the number of enabled synapse rows, and
add a raise statement in case it is not updated in a given run, to catch
a similar error more easily in the future.

Change-Id: I21d6b7c5ffd4d2bed54ee91ff578ed482f0bcda6