Drop hxv1 support in CI
With the introduction of hxv3, the surrounding software stack has
changed and the tag EOL_hxv1 no longer works with current haldls.
Therefore, we drop CI verification of v1 calibrations.

If a calibration for hxv1 is to be generated in the future, you may
cherry-pick commit e730be1ddacb6a065ad1a27cbe6899267142f441 for now, or
use an older version of haldls and below once there are further changes.

Change-Id: I95391dd7e12fd68218bba9547801a07a7b81b828