Improve calibration of hagen encoding
The hagen-mode encoding of activations into pulse lengths currently does
not work that well for low activations. The problem seems to be even
worse on v3. Here, we apply improvements to the calibration
methodology:

  * Use integration on synaptic input lines instead of neurons
  * Tweak DAC offsets for each driver individually while (before)
    calibrating STP ramp currents

The changes to the hagen-mode synapse driver calibration also
necessitate adjustments to the spiking (STP offset) calibration, using
much shared code. There, we also convert to integration on synaptic
input lines now.

Change-Id: I96433821bbb764199535f70898ddea0bf68c455d