Fix sporadically failing spiking neuron calib test
Sometimes, the fit while calibrating the membrane time constant has
failed in the spiking neuron calibration test without synaptic input
calib. This was caused by a high leak potential, which resulted in the
start of the exponential decay not being recognized properly, and
leaving too few samples for the fit.

Here, we fix the underlying issue of high leak potentials, which was
caused by an uncalibrated membrane time constant that resulted in too
little leakage for some neurons.

Also, we make the fit routine more robust to this behaviour and raise a
more descriptive error message if there are very few samples for the
fit.

Change-Id: Ic3222d6d76c4b8cec103e515a5d83ddddefd828f