Fix outliers in neuron threshold calib
Some neurons seem to be stuck at a low potential even though an offset
current should be flowing onto the membrane. They may be stuck in reset.
This resulted in higher threshold parameters than necessary, and at the
same time, lower measured threshold potentials than desired.

Performing a reset of the neurons before the threshold measurement seems
to fix this behaviour.

Since we should investigate this effect in more detail, issue 3996 was
created.

Change-Id: Ibef96224d5428cec29371313b7724cd757423470