Discard invalid initial MADC samples
Due to an issue that is not entirely understood yet, the MADC yields
broken samples for a few microseconds after a measurement is triggered.
See issue 4008 for details.

During MADC-based calibrations, we add a timeframe where samples are
recorded but discarded.

In the MADC recorder test, we now also assert the samples are sensible
and there are no unexpected edges in the trace, which would have been
the case previously.

Change-Id: Ife58fd36e399efb855fe6748ac9a98f1eaf7e868