Store calibration targets in dataclass
Instead of passing a multitude of keyword parameters to calibration
functions, we collect them in dataclasses.

They are split into targets, which mostly correspond to an observable,
and further options, that affect how the calibration is performed.

The previously used parameters are now marked deprecated and will be
removed in the future (issue 4017).

Change-Id: I4fc7edf1cf4834427b0d6a8f692f1d6ad5bc7719