Increase verbosity of hagen multiplication test
In case the hagen multiplication test fails with too much noise, we
print an array of the (average absolute value of) activations within
each run of the experiment.

Context:
This assertion fails sometimes in the nightly calibration runs. It seems
like sometimes, noise is _much_ higher than normally. We want to
investigate what happens in those runs, e.g., if there's one experiment
that generates a result which is far off the others.

Change-Id: If7b61cbd1dac03611a6d13eb563eb77a0d849df3