Add correlation sensors to spiking calib
A spiking calibration can now include the correlation sensors.
By default, they are not calibrated, as this takes a lot of time. But
when correlation_target is configured, they are calibrated, and the
calibration is merged with the neuron calib appropriately.

Note that the user still may have to change the connections to the
CADC in their experiment after loading the calibration. It is still
connected to the neuron readout, not reading correlation by default.

Change-Id: I5446eda613dc1d6975eeb9261416fcd7e353d626